Move LoadingSpinner component to new "common" directory
